[Bug #11003] - MPR application freezes for duration of volume creation after it's unloaded by memory manager  * ProgressGraphic performance improvements.   * Optimized a horridly inefficient progress graphic initialization routine.   * Draws are now short circuited even earlier if parent image is not visible in a tile.   * BackgroundTask progress sources are now throttled to update in 1 second intervals.  * Async imaging trial API updates.   * Async presentation images are now themselves a progress source, and also throttle updates to 1 second intervals.   * Async SOP frames no longer use Tasks to do background loading.    * The implementation must now provide the means of background loading.    * This allows them to potentially share the background thread with other objects.
--HG--
branch : Phoenix5
extra : rebase_source : 9932daa3b3ca897e1fc2e6e3b355f00aa9883541

+++ ImageViewer/Volume/Mpr/Trial/IAsyncSopFrameData.cs
@@ -32,7 +32,7 @@ namespace ClearCanvas.ImageViewer.Volume.Mpr
 	/// </remarks>
 	public interface IAsyncSopFrameData : ISopFrameData
 	{
-		float ProgressPercent { get; }
+		int ProgressPercent { get; }
 		event AsyncPixelDataProgressEventHandler ProgressChanged;
 		bool IsLoaded { get; }
 		event AsyncPixelDataEventHandler Loaded;
@@ -66,10 +66,10 @@ namespace ClearCanvas.ImageViewer.Volume.Mpr
 	/// </remarks>
 	public class AsyncPixelDataProgressEventArgs : AsyncPixelDataEventArgs
 	{
-		public float ProgressPercent { get; private set; }
+		public int ProgressPercent { get; private set; }
 		public bool IsComplete { get; private set; }
 
-		public AsyncPixelDataProgressEventArgs(float progressPercent, bool isComplete, byte[] pixelData)
+		public AsyncPixelDataProgressEventArgs(int progressPercent, bool isComplete, byte[] pixelData)
 			: base(pixelData)
 		{
 			ProgressPercent = progressPercent;
